+++ modules/user/usecase.go
@@ -8,6 +8,6 @@ import (
 type UserUsecase interface {
 	Create(ctx context.Context, user *models.User) error
 	Update(ctx context.Context, user *models.User) error
-	Get(ctx context.Context, id uint) (*models.User, error)
-	Delete(ctx context.Context, id uint) error
+	Get(ctx context.Context, id uint64) (*models.User, error)
+	Delete(ctx context.Context, id uint64) error
 }
